fix names and move user class to DAL assembly
