--- solution/Timebanks.NZ.DAL.MySql/Class1.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Timebanks.NZ.DAL.MySql
-{
-    public class Class1
-    {
-    }
-}
