+++ solution/Timebanks.NZ.DAL.MySql/TimebanksNZ.DAL.MySql.csproj
@@ -39,8 +39,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
+    <Compile Include="UserRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\TimebanksNZ.DAL\TimebanksNZ.DAL.csproj">
+      <Project>{8025838E-A7EA-4CE6-AFB4-FA13DE56F127}</Project>
+      <Name>TimebanksNZ.DAL</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
