+++ solution/TimebanksNZ.DAL/TimebanksNZ.DAL.csproj
@@ -40,6 +40,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Entities\IUser.cs" />
+    <Compile Include="Entities\User.cs" />
     <Compile Include="IRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
